grpc_fastapi/app/requirements.txt
fastapi 
uvicorn 
grpcio 
grpcio-tools
grpc-interceptor 
python-dotenv 
pydantic